Move yanny to extern
