+++ python/marvin/utils/general/maskbit.py
@@ -16,7 +16,7 @@ import numpy as np
 import pandas as pd
 
 import marvin
-from marvin.utils.general.yanny import yanny
+from marvin.extern.yanny import yanny
 
 
 # Stores the maskbits yanny file structure so that we don't need to open it more than once.
